fix: address review feedback — extract eventDescription, improve forcePushAllowed tests
- Extract eventDescription from ActivityFeed.tsx to pure event-description.ts
  to avoid motion/react transitive import in Jest test
- Re-export from ActivityFeed.tsx for backwards compatibility
- Update babysit-ui.test.ts to import from event-description.ts
- Replace trivial forcePushAllowed inline tests with tests exercising
  submitExternalPrToReviewQueue through MiniSql mock
- Enhance MiniSql mock to capture INSERT statements and params
